% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendario.R
\docType{data}
\name{calendario}
\alias{calendario}
\title{meses_curtos}
\format{
Vetor com os meses do ano:
\describe{
\item{ mes }{ numeric }
\item{ meses }{ character }
\item{ abreviacao }{ character }
\item{ abreviacao.us }{ character }
\item{ semestre }{ character }
\item{ trimestre }{ character }
\item{ bimestre }{ character }
}
}
\source{
Proprio.
}
\usage{
calendario
}
\description{
Relacao dos meses do ano de forma curta, em pt-Br.
}
\keyword{datasets}
