% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download data from the internet}
\usage{
download_data(url, csv_type = 2)
}
\arguments{
\item{url}{Endereço na internet onde o dado esta disponível}

\item{csv_type}{Type: 1 for "," separator
2 for ";" separator}
}
\value{
Tibble com os dados
}
\description{
Baixa dados da internet para um repositório local, de uma forma mais segura,
evitando a quebra de informações acessadas diretamente.
}
\examples{
url <- "https://docs.google.com/spreadsheets/d/e/<GOOGLE-SPREADSHEET-ID>
/pub?gid=<GID>&single=true&output=csv"
\dontrun{data <- download_data(url, csv_type = 1)}
}
