% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download data from the internet}
\usage{
download_data(url, csv_type = 2)
}
\arguments{
\item{url}{Endereço na internet onde o dado esta disponível}

\item{csv_type}{Type 1 for "," separator or  2 for ";" separator}
}
\value{
\code{Tibble} com os dados
}
\description{
Baixa dados da internet para um repositório local, de uma forma mais segura,
evitando a quebra de informações acessadas diretamente.
}
\examples{
url <- "https://myurl.com/data_file.csv"
\dontrun{
data <- download_data(url, csv_type = 1)
}
}
\seealso{
\link[utils]{download.file}
}
\author{
Herson Melo \href{mailto:hersonpc@gmail.com}{hersonpc@gmail.com}
}
